clean up volume indicator
